set up pokedex page structure
