--- .eslintrc
@@ -1,94 +1,0 @@
-{
-    "extends": "eslint:recommended",
-    "parserOptions": {
-        "ecmaVersion": 8,
-        "ecmaFeatures": {
-            "impliedStrict": true
-        },
-        "sourceType": "module"
-    },
-    "env": {
-        "browser": true,
-        "es6": true
-    },
-    "globals": {
-        "QUnit": false
-    },
-    "rules": {
-        "eqeqeq": [
-            "error",
-            "always"
-        ],
-        "no-console": "warn",
-        "indent": [
-            "error",
-            4,
-            {
-                "SwitchCase": 1,
-                "ignoreComments": true
-            }
-        ],
-        "quotes": [
-            "warn",
-            "single"
-        ],
-        "no-multi-spaces": [
-            "error",
-            {
-                "ignoreEOLComments": true
-            }
-        ],
-        "new-cap": "error",
-        "no-redeclare": [
-            "error",
-            {
-                "builtinGlobals": true
-            }
-        ],
-        "semi": [
-            "error",
-            "always"
-        ],
-        "space-in-parens": [
-            "error"
-        ],
-        "space-infix-ops": "error",
-        "object-curly-spacing": [
-            "error",
-            "always"
-        ],
-        "comma-spacing": "error",
-        "space-before-function-paren": [
-            "error",
-            "never"
-        ],
-        "keyword-spacing": [
-            "error",
-            {
-                "before": true,
-                "after": true,
-                "overrides": {
-                    "if": {
-                        "after": false
-                    },
-                    "for": {
-                        "after": false
-                    },
-                    "while": {
-                        "after": false
-                    },
-                    "do": {
-                        "after": false
-                    },
-                    "switch": {
-                        "after": false
-                    },
-                    "catch": {
-                        "after": false
-                    }
-                }
-            }
-        ],
-        "array-bracket-spacing": "error"
-    }
-}
