+++ index.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <link rel="stylesheet" href="./src/main.css">
+  <title>Document</title>
+</head>
+<body>
+  <header>
+    <h1></h1>
+  </header>
+  <main>
+
+  </main>
+
+  <footer></footer>
+  <script type="module" src="./src/index.js"></script>
+</body>
+</html>
