--- test/tests.js
@@ -1,5 +1,0 @@
-import './html-equal.js';
-
-QUnit.test('delete me and write a test file', assert => {
-    assert.equal(true, true);
-});
