+++ tests/index.js
@@ -1,6 +1,4 @@
 /* eslint-disable */
-
-// include jsdom for DOM use in tests on travis
 const jsdom = require("jsdom");
 const { JSDOM } = jsdom;
 const { window } = new JSDOM(``);
