+++ tests/tests.js
@@ -1,0 +1,2 @@
+import './html-equal.js';
+// import some tests!
